Users are now able to review specific theaters and view specific theaters' reviews

+++ MoviesForEveryone/Models/Theater.cs
@@ -11,11 +11,11 @@ namespace MoviesForEveryone.Models
         [Key]
         public int Id { get; set; }
         public string theaterName { get; set; }
-        public float avgClean { get; set; }        
-        public float avgConc { get; set; }
-        public float avgArcade { get; set; }
-        public float avgViewing { get; set; }
-        public float totalAvg { get; set; }
+        public double avgClean { get; set; }        
+        public double avgConc { get; set; }
+        public double avgArcade { get; set; }
+        public double avgViewing { get; set; }
+        public double totalAvg { get; set; }
         public void recalcAvg()
         {
             totalAvg = (avgClean + avgConc + avgArcade + avgViewing + totalAvg) / 5;
